add %%CURRENT_PAGE%% magic placeholder
Useful for default page templates (like `<catlist %%CURRENT_PAGE%% -noHead -sortAscending -hideNotFoundMsg>`)

When a page is created by default nothing will be shown with that catlist. When creating subpages (without moving the original page into the new namespace and renaming it to start) they will be shown on the original page as links automatically.

I'm not sure what the difference between `getID()` and `global $ID` is, but I thought it would be better style adding the variable here instead of in the `handle` function.